Editing a View in SQLite
Editing a view in SQLlite first requires you to drop the table in order to edit it. I provided a screenshot where I modify the view, but when modifying the view a display message will show the current view needs to be dropped. I made a slight change to the view I created and ran the query. The change I made is shown in the screenshot.